+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Unreleased
 
+* Add label size option for select ([PR #1131](https://github.com/alphagov/govuk_publishing_components/pull/1131))
 * Update search component styles [PR #1128](https://github.com/alphagov/govuk_publishing_components/pull/1128)
 * Add name option to select component ([PR #1130](https://github.com/alphagov/govuk_publishing_components/pull/1130))